+++ allauth/account/tests.py
@@ -1,11 +1,11 @@
 from __future__ import absolute_import
 
 import json
-import unittest
 import uuid
 from datetime import timedelta
 
 import django
+from django import forms
 from django.conf import settings
 from django.contrib.auth.models import AbstractUser, AnonymousUser
 from django.contrib.sites.models import Site
@@ -16,7 +16,7 @@ from django.test.client import Client, RequestFactory
 from django.test.utils import override_settings
 from django.utils.timezone import now
 
-from allauth.account.forms import BaseSignupForm
+from allauth.account.forms import BaseSignupForm, SignupForm
 from allauth.account.models import (
     EmailAddress,
     EmailConfirmation,
@@ -70,42 +70,42 @@ class AccountTests(TestCase):
         user.set_password('psst')
         user.save()
         EmailAddress.objects.create(user=user,
-                                    email='raymond.penners@gmail.com',
+                                    email="raymond.penners@example.com",
                                     primary=True,
                                     verified=True)
         resp = self.client.post(reverse('account_login'),
                                 {'login': '@raymond.penners',
                                  'password': 'psst'})
         self.assertRedirects(resp,
-                             'http://testserver' + settings.LOGIN_REDIRECT_URL,
+                             settings.LOGIN_REDIRECT_URL,
                              fetch_redirect_response=False)
 
     def test_signup_same_email_verified_externally(self):
-        user = self._test_signup_email_verified_externally('john@doe.com',
-                                                           'john@doe.com')
+        user = self._test_signup_email_verified_externally("john@example.com",
+                                                           "john@example.com")
         self.assertEqual(EmailAddress.objects.filter(user=user).count(),
                          1)
         EmailAddress.objects.get(verified=True,
-                                 email='john@doe.com',
+                                 email="john@example.com",
                                  user=user,
                                  primary=True)
 
     def test_signup_other_email_verified_externally(self):
         """
-        John is invited on john@work.com, but signs up via john@home.com.
+        John is invited on john@example.org, but signs up via john@example.com.
         E-mail verification is by-passed, their home e-mail address is
         used as a secondary.
         """
-        user = self._test_signup_email_verified_externally('john@home.com',
-                                                           'john@work.com')
+        user = self._test_signup_email_verified_externally("john@example.com",
+                                                           "john@example.org")
         self.assertEqual(EmailAddress.objects.filter(user=user).count(),
                          2)
         EmailAddress.objects.get(verified=False,
-                                 email='john@home.com',
+                                 email="john@example.com",
                                  user=user,
                                  primary=False)
         EmailAddress.objects.get(verified=True,
-                                 email='john@work.com',
+                                 email="john@example.org",
                                  user=user,
                                  primary=True)
 
@@ -140,7 +140,7 @@ class AccountTests(TestCase):
             reverse('account_signup'),
             data={
                 'username': 'johndoe',
-                'email': 'john@work.com',
+                'email': 'john@example.org',
                 'password1': 'johndoe',
                 'password2': 'janedoe'})
         self.assertFormError(
@@ -156,8 +156,8 @@ class AccountTests(TestCase):
     def test_signup_email_twice(self):
         request = RequestFactory().post(reverse('account_signup'),
                                         {'username': 'johndoe',
-                                         'email': 'john@work.com',
-                                         'email2': 'john@work.com',
+                                         'email': 'john@example.org',
+                                         'email2': 'john@example.org',
                                          'password1': 'johndoe',
                                          'password2': 'johndoe'})
         from django.contrib.messages.middleware import MessageMiddleware
@@ -168,7 +168,7 @@ class AccountTests(TestCase):
         from .views import signup
         signup(request)
         user = get_user_model().objects.get(username='johndoe')
-        self.assertEqual(user.email, 'john@work.com')
+        self.assertEqual(user.email, "john@example.org")
 
     def _create_user(self, username='john', password='doe'):
         user = get_user_model().objects.create(
@@ -191,8 +191,9 @@ class AccountTests(TestCase):
         self._create_user_and_login()
         c = self.client
         resp = c.get(reverse('account_login'))
-        self.assertRedirects(resp, 'http://testserver/accounts/profile/',
-                             fetch_redirect_response=False)
+        self.assertRedirects(
+            resp, "/accounts/profile/", fetch_redirect_response=False
+        )
 
     def test_password_reset_get(self):
         resp = self.client.get(reverse('account_reset_password'))
@@ -259,14 +260,14 @@ class AccountTests(TestCase):
 
     def _request_new_password(self):
         user = get_user_model().objects.create(
-            username='john', email='john@doe.org', is_active=True)
+            username='john', email="john@example.org", is_active=True)
         user.set_password('doe')
         user.save()
         self.client.post(
             reverse('account_reset_password'),
-            data={'email': 'john@doe.org'})
+            data={'email': 'john@example.org'})
         self.assertEqual(len(mail.outbox), 1)
-        self.assertEqual(mail.outbox[0].to, ['john@doe.org'])
+        self.assertEqual(mail.outbox[0].to, ["john@example.org"])
         return user
 
     def test_password_reset_flow(self):
@@ -349,12 +350,12 @@ class AccountTests(TestCase):
         # Signup
         resp = c.post(reverse('account_signup'),
                       {'username': 'johndoe',
-                       'email': 'john@doe.com',
+                       'email': 'john@example.com',
                        'password1': 'johndoe',
                        'password2': 'johndoe'},
                       follow=True)
         self.assertEqual(resp.status_code, 200)
-        self.assertEqual(mail.outbox[0].to, ['john@doe.com'])
+        self.assertEqual(mail.outbox[0].to, ['john@example.com'])
         self.assertGreater(mail.outbox[0].body.find('https://'), 0)
         self.assertEqual(len(mail.outbox), 1)
         self.assertTemplateUsed(
@@ -380,7 +381,7 @@ class AccountTests(TestCase):
             self.assertEqual(len(mail.outbox), attempt)
             self.assertEqual(
                 EmailConfirmation.objects.filter(
-                    email_address__email='john@doe.com').count(),
+                    email_address__email='john@example.com').count(),
                 attempt)
             # Wait for cooldown
             EmailConfirmation.objects.update(sent=now() - timedelta(days=1))
@@ -400,7 +401,7 @@ class AccountTests(TestCase):
                       {'login': 'johndoe',
                        'password': 'johndoe'})
         self.assertRedirects(resp,
-                             'http://testserver' + settings.LOGIN_REDIRECT_URL,
+                             settings.LOGIN_REDIRECT_URL,
                              fetch_redirect_response=False)
 
     def test_email_escaping(self):
@@ -409,7 +410,7 @@ class AccountTests(TestCase):
         site.save()
         u = get_user_model().objects.create(
             username='test',
-            email='foo@bar.com')
+            email="user@example.com")
         request = RequestFactory().get('/')
         EmailAddress.objects.add_email(request, u, u.email, confirm=True)
         self.assertTrue(mail.outbox[0].subject[1:].startswith(site.name))
@@ -423,14 +424,14 @@ class AccountTests(TestCase):
         user.set_password('doe')
         user.save()
         EmailAddress.objects.create(user=user,
-                                    email='john@example.com',
+                                    email="user@example.com",
                                     primary=True,
                                     verified=False)
         resp = self.client.post(reverse('account_login'),
                                 {'login': 'john',
                                  'password': 'doe'})
         self.assertRedirects(resp,
-                             'http://testserver' + settings.LOGIN_REDIRECT_URL,
+                             settings.LOGIN_REDIRECT_URL,
                              fetch_redirect_response=False)
 
     @override_settings(
@@ -442,7 +443,7 @@ class AccountTests(TestCase):
         user.set_password('doe')
         user.save()
         EmailAddress.objects.create(user=user,
-                                    email='john@example.com',
+                                    email="user@example.com",
                                     primary=True,
                                     verified=False)
         for i in range(5):
@@ -469,7 +470,7 @@ class AccountTests(TestCase):
         user.set_password('doe')
         user.save()
         EmailAddress.objects.create(user=user,
-                                    email='john@example.com',
+                                    email="user@example.com",
                                     primary=True,
                                     verified=False)
         resp = self.client.post(reverse('account_login'),
@@ -503,7 +504,7 @@ class AccountTests(TestCase):
         user.set_password('john')
         user.save()
         EmailAddress.objects.create(user=user,
-                                    email='doe@example.com',
+                                    email="user@example.com",
                                     primary=True,
                                     verified=False)
         resp = self.client.post(reverse('account_login'),
@@ -513,13 +514,13 @@ class AccountTests(TestCase):
 
     def test_ajax_password_reset(self):
         get_user_model().objects.create(
-            username='john', email='john@doe.org', is_active=True)
+            username='john', email="john@example.org", is_active=True)
         resp = self.client.post(
             reverse('account_reset_password'),
-            data={'email': 'john@doe.org'},
+            data={'email': 'john@example.org'},
             HTTP_X_REQUESTED_WITH='XMLHttpRequest')
         self.assertEqual(len(mail.outbox), 1)
-        self.assertEqual(mail.outbox[0].to, ['john@doe.org'])
+        self.assertEqual(mail.outbox[0].to, ["john@example.org"])
         self.assertEqual(resp['content-type'], 'application/json')
 
     def test_ajax_login_fail(self):
@@ -581,14 +582,14 @@ class AccountTests(TestCase):
         c.get(reverse('account_signup'))
         resp = c.post(reverse('account_signup'),
                       {'username': 'johndoe',
-                       'email': 'john@doe.com',
+                       'email': 'john@example.com',
                        'password1': 'johndoe',
                        'password2': 'johndoe'})
         # Logged in
         self.assertRedirects(resp,
                              settings.LOGIN_REDIRECT_URL,
                              fetch_redirect_response=False)
-        self.assertEqual(mail.outbox[0].to, ['john@doe.com'])
+        self.assertEqual(mail.outbox[0].to, ['john@example.com'])
         self.assertEqual(len(mail.outbox), 1)
         # Logout & login again
         c.logout()
@@ -601,7 +602,7 @@ class AccountTests(TestCase):
         self.assertRedirects(resp,
                              settings.LOGIN_REDIRECT_URL,
                              fetch_redirect_response=False)
-        self.assertEqual(mail.outbox[0].to, ['john@doe.com'])
+        self.assertEqual(mail.outbox[0].to, ['john@example.com'])
         # There was an issue that we sent out email confirmation mails
         # on each login in case of optional verification. Make sure
         # this is not the case:
@@ -626,7 +627,7 @@ class AccountTests(TestCase):
         resp = self.client.post(
             reverse('account_signup'),
             {'username': 'johndoe',
-             'email': 'john@doe.com',
+             'email': 'john@example.com',
              'password1': 'johndoe',
              'password2': 'johndoe'})
         self.assertFormError(resp, 'form', None, [])
@@ -705,7 +706,7 @@ class EmailFormTests(TestCase):
     def setUp(self):
         User = get_user_model()
         self.user = User.objects.create(username='john',
-                                        email='john1@doe.org')
+                                        email="john1@example.org")
         self.user.set_password('doe')
         self.user.save()
         self.email_address = EmailAddress.objects.create(
@@ -715,7 +716,7 @@ class EmailFormTests(TestCase):
             primary=True)
         self.email_address2 = EmailAddress.objects.create(
             user=self.user,
-            email='john2@doe.org',
+            email="john2@example.org",
             verified=False,
             primary=False)
         self.client.login(username='john', password='doe')
@@ -724,9 +725,9 @@ class EmailFormTests(TestCase):
         resp = self.client.post(
             reverse('account_email'),
             {'action_add': '',
-             'email': 'john3@doe.org'})
+             'email': 'john3@example.org'})
         EmailAddress.objects.get(
-            email='john3@doe.org',
+            email="john3@example.org",
             user=self.user,
             verified=False,
             primary=False)
@@ -739,10 +740,10 @@ class EmailFormTests(TestCase):
             HTTP_X_REQUESTED_WITH='XMLHttpRequest')
         data = json.loads(resp.content.decode('utf8'))
         assert data['data'] == [
-            {'email': 'john1@doe.org',
+            {'email': 'john1@example.org',
              'primary': True,
              'verified': True},
-            {'email': 'john2@doe.org',
+            {'email': 'john2@example.org',
              'primary': False,
              'verified': False}]
 
@@ -750,7 +751,7 @@ class EmailFormTests(TestCase):
         resp = self.client.post(
             reverse('account_email'),
             {'action_add': '',
-             'email': 'john3@doe.org'},
+             'email': 'john3@example.org'},
             HTTP_X_REQUESTED_WITH='XMLHttpRequest')
         data = json.loads(resp.content.decode('utf8'))
         self.assertEqual(data['location'],
@@ -760,7 +761,7 @@ class EmailFormTests(TestCase):
         resp = self.client.post(
             reverse('account_email'),
             {'action_add': '',
-             'email': 'john3#doe.org'},
+             'email': 'john3#example.org'},
             HTTP_X_REQUESTED_WITH='XMLHttpRequest')
         data = json.loads(resp.content.decode('utf8'))
         assert 'valid' in data['form']['fields']['email']['errors'][0]
@@ -900,12 +901,45 @@ class BaseSignupFormTests(TestCase):
         self.assertFalse(form.is_valid())
 
 
+class CustomSignupFormTests(TestCase):
+
+    @override_settings(
+        ACCOUNT_SIGNUP_EMAIL_ENTER_TWICE=True,
+        ACCOUNT_SIGNUP_PASSOWRD_ENTER_TWICE=True)
+    def test_custom_form_field_order(self):
+
+        expected_field_order = [
+            'email',
+            'email2',
+            'password1',
+            'password2',
+            'username',
+            'last_name',
+            'first_name'
+        ]
+
+        class TestSignupForm(forms.Form):
+            first_name = forms.CharField(max_length=30)
+            last_name = forms.CharField(max_length=30)
+
+            field_order = expected_field_order
+
+        class CustomSignupForm(SignupForm, TestSignupForm):
+            # ACCOUNT_SIGNUP_FORM_CLASS is only abided by when the
+            # BaseSignupForm definition is loaded the first time on Django
+            # startup. @override_settings() has therefore no effect.
+            pass
+
+        form = CustomSignupForm()
+        self.assertEqual(list(form.fields.keys()), expected_field_order)
+
+
 class AuthenticationBackendTests(TestCase):
 
     def setUp(self):
         user = get_user_model().objects.create(
             is_active=True,
-            email='john@doe.com',
+            email='john@example.com',
             username='john')
         user.set_password(user.username)
         user.save()
@@ -960,42 +994,37 @@ class AuthenticationBackendTests(TestCase):
             user.pk)
 
 
+class UUIDUser(AbstractUser):
+    id = models.UUIDField(
+        primary_key=True, default=uuid.uuid4, editable=False
+    )
+
+    class Meta(AbstractUser.Meta):
+        swappable = "AUTH_USER_MODEL"
+
+
 class UtilsTests(TestCase):
     def setUp(self):
-        if hasattr(models, 'UUIDField'):
-            self.user_id = uuid.uuid4().hex
+        self.user_id = uuid.uuid4().hex
 
-            class UUIDUser(AbstractUser):
-                id = models.UUIDField(primary_key=True,
-                                      default=uuid.uuid4,
-                                      editable=False)
-
-                class Meta(AbstractUser.Meta):
-                    swappable = 'AUTH_USER_MODEL'
-        else:
-            UUIDUser = get_user_model()
-        self.UUIDUser = UUIDUser
-
-    @unittest.skipUnless(hasattr(models, 'UUIDField'),
-                         reason="No UUIDField in this django version")
     def test_url_str_to_pk_identifies_UUID_as_stringlike(self):
         with patch('allauth.account.utils.get_user_model') as mocked_gum:
-            mocked_gum.return_value = self.UUIDUser
+            mocked_gum.return_value = UUIDUser
             self.assertEqual(url_str_to_user_pk(self.user_id),
                              self.user_id)
 
     def test_pk_to_url_string_identifies_UUID_as_stringlike(self):
-        user = self.UUIDUser(
+        user = UUIDUser(
             is_active=True,
-            email='john@doe.com',
+            email='john@example.com',
             username='john')
-        self.assertEquals(user_pk_to_url_str(user), str(user.pk))
+        self.assertEqual(user_pk_to_url_str(user), str(user.pk))
 
     @override_settings(ACCOUNT_PRESERVE_USERNAME_CASING=False)
     def test_username_lower_cased(self):
         user = get_user_model()()
         user_username(user, 'CamelCase')
-        self.assertEquals(user_username(user), 'camelcase')
+        self.assertEqual(user_username(user), 'camelcase')
         # TODO: Actually test something
         filter_users_by_username('CamelCase', 'FooBar')
 
@@ -1003,6 +1032,6 @@ class UtilsTests(TestCase):
     def test_username_case_preserved(self):
         user = get_user_model()()
         user_username(user, 'CamelCase')
-        self.assertEquals(user_username(user), 'CamelCase')
+        self.assertEqual(user_username(user), 'CamelCase')
         # TODO: Actually test something
         filter_users_by_username('camelcase', 'foobar')