+++ allauth/account/views.py
@@ -231,11 +231,12 @@ class SignupView(RedirectAuthenticatedUserMixin, CloseableSignupMixin,
         ret = super(SignupView, self).get_context_data(**kwargs)
         form = ret['form']
         email = self.request.session.get('account_verified_email')
-        email_keys = ['email']
-        if app_settings.SIGNUP_EMAIL_ENTER_TWICE:
-            email_keys.append('email2')
-        for email_key in email_keys:
-            form.fields[email_key].initial = email
+        if email:
+            email_keys = ['email']
+            if app_settings.SIGNUP_EMAIL_ENTER_TWICE:
+                email_keys.append('email2')
+            for email_key in email_keys:
+                form.fields[email_key].initial = email
         login_url = passthrough_next_redirect_url(self.request,
                                                   reverse("account_login"),
                                                   self.redirect_field_name)