+++ allauth/socialaccount/providers/mailchimp/tests.py
@@ -19,11 +19,11 @@ class MailChimpTests(OAuth2TestsMixin, TestCase):
             "accountname": "Name can have spaces",
             "user_id": "99999999",
             "login": {
-                "email": "test@test.com",
+                "email": "test@example.com",
                 "avatar": "http://gallery.mailchimp.com/1a1a/avatar/2a2a.png",
                 "login_id": "88888888",
-                "login_name": "test@test.com",
-                "login_email": "test@test.com"
+                "login_name": "test@example.com",
+                "login_email": "test@example.com"
             },
             "login_url": "https://login.mailchimp.com",
             "api_endpoint": "https://usX.api.mailchimp.com"