+++ allauth/socialaccount/providers/persona/tests.py
@@ -17,11 +17,11 @@ class PersonaTests(TestCase):
                    '.requests') as requests_mock:
             requests_mock.post.return_value.json.return_value = {
                 'status': 'okay',
-                'email': 'persona@mail.com'
+                'email': 'persona@example.com'
             }
 
             resp = self.client.post(reverse('persona_login'),
                                     dict(assertion='dummy'))
             self.assertRedirects(resp, '/accounts/profile/',
                                  fetch_redirect_response=False)
-            get_user_model().objects.get(email='persona@mail.com')
+            get_user_model().objects.get(email='persona@example.com')